Adding the initial cycling through of Vulns
Still need to add the CCI to NIST into memory to cycle
through, but this lets me get the model in my head down
of oging through each checklist based on system and match to
the CCI and NIST. We need to do a lookup on the NISTControls
listing still and do an add on the ComplianceRecord to the
control for what we find if there. Otherwise make a new one.